Add Wrap and Wrapf functions
Currently github.com/pkg/errors.Wrap/Wrapf functions
overwrite already attached stack trace.

From a UX/DX point of view one would like to add
stack trace to an error if none is attached yet.

There are several open issues discussing the problem in the original repo:

https://github.com/pkg/errors/issues/75
https://github.com/pkg/errors/issues/144
https://github.com/pkg/errors/issues/158

At the moment there is no solution provided,
but it looks like the author is willing to make some changes.
Until then this commit adds custom Wrap/Wrapf functions
which implement this desired behaviour.
Other than that, they behave the same way.

There is also a pending PR in the original repo:

https://github.com/pkg/errors/pull/122

It worths mentioning that there are alternative solutions,
like merging stack traces:

https://github.com/srvc/fail/pull/13

After examining the solution it seems that it's
probably too complicated and unnecessary for most cases.
It's also unlikely that the original errors package
will implement this behaviour, so for now
we stick to the most expected behaviour.